# Restructure compare_greater function used in parquet statistics for better performance